bcftbx.TabFile: relocate TabFileIterator from cli.best_exons.
